zadanie: testy jednostkowe

+++ kodilla-basic-tests/src/main/java/com/kodilla/bank/homework/Bank.java
@@ -7,8 +7,8 @@ public class Bank {
     public Bank(CashMachine[] cashMachines) {
             this.cashMachines = cashMachines;
     }
-    public double getSaldoOfAllCashMachine() {
-        double totalSaldo = 0;
+    public int getSaldoOfAllCashMachine() {
+        int totalSaldo = 0;
         for (CashMachine machine : cashMachines) {
             totalSaldo += machine.getSaldo();
         }
